+++ src/Container.php
@@ -47,7 +47,8 @@ use Infection\Configuration\Schema\SchemaConfigurationLoader;
 use Infection\Configuration\Schema\SchemaValidator;
 use Infection\Differ\DiffColorizer;
 use Infection\Differ\Differ;
-use Infection\Event\EventDispatcher;
+use Infection\Event\EventDispatcher\EventDispatcher;
+use Infection\Event\EventDispatcher\SyncEventDispatcher;
 use Infection\FileSystem\SourceFileCollector;
 use Infection\FileSystem\TmpDirProvider;
 use Infection\Locator\RootsFileLocator;
@@ -193,8 +194,8 @@ final class Container
             Differ::class => static function (): Differ {
                 return new Differ(new BaseDiffer());
             },
-            EventDispatcher::class => static function (): EventDispatcher {
-                return new EventDispatcher();
+            SyncEventDispatcher::class => static function (): SyncEventDispatcher {
+                return new SyncEventDispatcher();
             },
             ParallelProcessRunner::class => static function (self $container): ParallelProcessRunner {
                 return new ParallelProcessRunner($container->getEventDispatcher());
@@ -551,7 +552,7 @@ final class Container
 
     public function getEventDispatcher(): EventDispatcher
     {
-        return $this->get(EventDispatcher::class);
+        return $this->get(SyncEventDispatcher::class);
     }
 
     public function getParallelProcessRunner(): ParallelProcessRunner
