+++ tests/phpunit/Event/Subscriber/CiMutantCreatingConsoleLoggerSubscriberTest.php
@@ -35,7 +35,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Event\Subscriber;
 
-use Infection\Event\EventDispatcher;
+use Infection\Event\EventDispatcher\SyncEventDispatcher;
 use Infection\Event\MutantsCreatingStarted;
 use Infection\Event\Subscriber\CiMutantCreatingConsoleLoggerSubscriber;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -65,7 +65,7 @@ final class CiMutantCreatingConsoleLoggerSubscriberTest extends TestCase
                 'Creating mutated files and processes: 123',
             ]);
 
-        $dispatcher = new EventDispatcher();
+        $dispatcher = new SyncEventDispatcher();
         $dispatcher->addSubscriber(new CiMutantCreatingConsoleLoggerSubscriber($this->output));
 
         $dispatcher->dispatch(new MutantsCreatingStarted(123));
