+++ tests/phpunit/Process/Runner/Parallel/ParallelProcessRunnerTest.php
@@ -35,7 +35,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\Process\Runner\Parallel;
 
-use Infection\Event\EventDispatcher;
+use Infection\Event\EventDispatcher\EventDispatcher;
 use Infection\Event\MutantProcessFinished;
 use Infection\Mutant\Mutant;
 use Infection\Process\MutantProcess;
